refactor(mesh): rename gql query to singular
